fixed a bug when writing large numbes in LCD

+++ ARM System/02-HAL/01-HLCD/HLCD_interface.h
@@ -8,7 +8,7 @@ void HLCD_voidInit(void);
 void HLCD_voidWriteCMD(u8 cmd);
 void HLCD_voidWriteData(u8 data);
 void HLCD_voidWriteStr(u8 str[]);
-void HLCD_voidWriteNumber(u16 n);
+void HLCD_voidWriteNumber(u32 n);
 void HLCD_voidSetCursor(u8 row, u8 col);
 void HLCD_voidAddCustomCharacter(u8 indx, u8 pattern[7]);
 void HLCD_voidWriteCustomCharacter(u8 indx);
